Move PosixSourceAccessor into its own file
