+++ src/libutil/source-accessor.hh
@@ -104,31 +104,4 @@ struct SourceAccessor
     virtual std::string showPath(const CanonPath & path);
 };
 
-/**
- * A source accessor that uses the Unix filesystem.
- */
-struct PosixSourceAccessor : SourceAccessor
-{
-    /**
-     * The most recent mtime seen by lstat(). This is a hack to
-     * support dumpPathAndGetMtime(). Should remove this eventually.
-     */
-    time_t mtime = 0;
-
-    void readFile(
-        const CanonPath & path,
-        Sink & sink,
-        std::function<void(uint64_t)> sizeCallback) override;
-
-    bool pathExists(const CanonPath & path) override;
-
-    Stat lstat(const CanonPath & path) override;
-
-    DirEntries readDirectory(const CanonPath & path) override;
-
-    std::string readLink(const CanonPath & path) override;
-
-    std::optional<CanonPath> getPhysicalPath(const CanonPath & path) override;
-};
-
 }
